requirements.txt
pytest==8.4.1
selenium==4.35.0
webdriver-manager==4.0.2
allure-pytest==2.15.0
python-dotenv==1.2.1